Added more tests for answers.

+++ src/mname.h
@@ -146,9 +146,9 @@ int mname_get_a_rdata(
 int mname_get_offset( const struct mname_msg* msg_in, size_t msg_buf_sz,
    uint16_t idx );
 int mname_add_answer( struct mname_msg* msg, size_t msg_buf_sz,
-   char* domain, size_t domain_len,
+   char* domain, uint8_t domain_len,
    uint16_t type, uint16_t class, uint32_t ttl,
-   char* rdata, size_t rdata_len );
+   char* rdata, uint16_t rdata_len );
 
 #endif /* MNAME_H */
